Add installation scripts.
