+++ torchbenchmark/models/detectron2_fasterrcnn_r_101_c4/install.py
