+++ torchbenchmark/models/detectron2_fasterrcnn_r_101_dc5/install.py
