+++ torchbenchmark/models/detectron2_fasterrcnn_r_101_fpn/install.py
