+++ torchbenchmark/models/detectron2_fasterrcnn_r_50_c4/install.py
