+++ torchbenchmark/models/detectron2_fasterrcnn_r_50_dc5/install.py
