+++ torchbenchmark/models/detectron2_fasterrcnn_r_50_fpn/install.py
