--- torchbenchmark/models/detectron2_maskrcnn/requirements.txt
@@ -1,3 +1,0 @@
-git+https://github.com/facebookresearch/detectron2.git@1fbb4b3
-omegaconf==2.1.1
-numpy
