+++ torchbenchmark/models/detectron2_maskrcnn_r_50_fpn/install.py
