+++ torchbenchmark/util/framework/detectron2/__init__.py
@@ -1,121 +1,44 @@
-from torchbenchmark.util.model import BenchmarkModel
-import itertools
 import os
+import sys
+import subprocess
 from pathlib import Path
-import torch
+from urllib import request
 
-# setup environment variable
 CURRENT_DIR = Path(os.path.dirname(os.path.realpath(__file__)))
-DATA_DIR = os.path.join(CURRENT_DIR.parent.parent.parent, "data", ".data", "coco2017-minimal")
-assert os.path.exists(DATA_DIR), "Couldn't find coco2017 minimal data dir, please run install.py again."
-if not 'DETECTRON2_DATASETS' in os.environ:
-    os.environ['DETECTRON2_DATASETS'] = DATA_DIR
+# Load pre-trained weights
+# copied from https://github.com/facebookresearch/detectron2/blob/5934a1452801e669bbf9479ae222ce1a8a51f52e/MODEL_ZOO.md
+MODEL_WEIGHTS_MAP = {
+    "detectron2_fasterrcnn_r_50_c4": "https://dl.fbaipublicfiles.com/detectron2/COCO-Detection/faster_rcnn_R_50_C4_1x/137257644/model_final_721ade.pkl",
+    "detectron2_fasterrcnn_r_50_dc5": "https://dl.fbaipublicfiles.com/detectron2/COCO-Detection/faster_rcnn_R_50_DC5_1x/137847829/model_final_51d356.pkl",
+    "detectron2_fasterrcnn_r_50_fpn": "https://dl.fbaipublicfiles.com/detectron2/COCO-Detection/faster_rcnn_R_50_FPN_1x/137257794/model_final_b275ba.pkl",
+    "detectron2_fasterrcnn_r_101_c4": "https://dl.fbaipublicfiles.com/detectron2/COCO-Detection/faster_rcnn_R_101_C4_3x/138204752/model_final_298dad.pkl",
+    "detectron2_fasterrcnn_r_101_dc5": "https://dl.fbaipublicfiles.com/detectron2/COCO-Detection/faster_rcnn_R_101_DC5_3x/138204841/model_final_3e0943.pkl",
+    "detectron2_fasterrcnn_r_101_fpn": "https://dl.fbaipublicfiles.com/detectron2/COCO-Detection/faster_rcnn_R_101_FPN_3x/137851257/model_final_f6e8b1.pkl",
+    "detectron2_maskrcnn_r_50_c4": "https://dl.fbaipublicfiles.com/detectron2/COCO-InstanceSegmentation/mask_rcnn_R_50_C4_1x/137259246/model_final_9243eb.pkl",
+    "detectron2_maskrcnn_r_50_fpn": "https://dl.fbaipublicfiles.com/detectron2/COCO-InstanceSegmentation/mask_rcnn_R_50_FPN_1x/137260431/model_final_a54504.pkl",
+    "detectron2_maskrcnn_r_101_c4": "https://dl.fbaipublicfiles.com/detectron2/COCO-InstanceSegmentation/mask_rcnn_R_101_C4_3x/138363239/model_final_a2914c.pkl",
+    "detectron2_maskrcnn_r_101_fpn": "https://dl.fbaipublicfiles.com/detectron2/COCO-InstanceSegmentation/mask_rcnn_R_101_FPN_3x/138205316/model_final_a3ec72.pkl",
+    "detectron2_maskrcnn": "https://dl.fbaipublicfiles.com/detectron2/COCO-InstanceSegmentation/mask_rcnn_R_50_FPN_1x/137260431/model_final_a54504.pkl",
+}
 
-from detectron2.modeling.meta_arch.build import META_ARCH_REGISTRY
-from detectron2.engine import default_argument_parser
-from detectron2.solver import build_optimizer
-from detectron2.config import LazyConfig, get_cfg, instantiate
-from detectron2 import model_zoo
-from detectron2.utils.events import EventStorage
-from torch.utils._pytree import tree_map, tree_flatten
+def check_data_dir():
+    coco2017_data_dir = os.path.join(CURRENT_DIR.parent.parent.parent, "data", ".data", "coco2017-minimal")
+    assert os.path.exists(coco2017_data_dir), "Couldn't find coco2017 minimal data dir, please run install.py again."
 
-from typing import Tuple
+def install_model_weights(model_name, model_dir):
+    assert model_name in MODEL_WEIGHTS_MAP, f"Model {model_name} is not in MODEL_WEIGHTS_MAP. Cannot download the model weights file."
+    model_full_path = os.path.join(model_dir, ".data", f"{model_name}.pkl")
+    # download the file if not exists
+    # TODO: verify the model file integrity
+    if os.path.exists(model_full_path):
+        return
+    request.urlretrieve(MODEL_WEIGHTS_MAP[model_name], model_full_path)
 
-def setup(args):
-    if args.config_file.endswith(".yaml"):
-        cfg = get_cfg()
-        cfg.merge_from_file(args.config_file)
-        cfg.SOLVER.BASE_LR = 0.001  # Avoid NaNs. Not useful in this script anyway.
-        # set images per batch to 1
-        cfg.SOLVER.IMS_PER_BATCH = 1
-        cfg.merge_from_list(args.opts)
-        cfg.freeze()
-    else:
-        cfg = LazyConfig.load(args.config_file)
-        cfg = LazyConfig.apply_overrides(cfg, args.opts)
-    return cfg
+def pip_install_requirements():
+    requirements_file = os.path.join(CURRENT_DIR, "requirements.txt")
+    subprocess.check_call([sys.executable, '-m', 'pip', 'install', '-q', '-r', requirements_file])
 
-def build_model(cfg):
-    """
-    Build the whole model architecture, defined by ``cfg.MODEL.META_ARCHITECTURE``.
-    Note that it does not load any weights from ``cfg``.
-    """
-    meta_arch = cfg.MODEL.META_ARCHITECTURE
-    model = META_ARCH_REGISTRY.get(meta_arch)(cfg)
-    return model
-
-def prefetch(dataloader, device, precision="fp32"):
-    r = []
-    dtype = torch.float16 if precision == "fp16" else torch.float32
-    for batch in dataloader:
-        r.append(tree_map(lambda x: x.to(device, dtype=dtype) if isinstance(x, torch.Tensor) else x, batch))
-    return r
-
-def get_abs_path(config):
-    import detectron2
-    detectron2_root = os.path.abspath(os.path.dirname(detectron2.__file__))
-    return os.path.join(detectron2_root, "model_zoo", "configs", config)
-
-class Detectron2Model(BenchmarkModel):
-    # To recognize this is a detectron2 model
-    DETECTRON2_MODEL = True
-    # Default eval precision on CUDA device is fp16
-    DEFAULT_EVAL_CUDA_PRECISION = "fp16"
-
-    def __init__(self, variant, test, device, jit=False, batch_size=None, extra_args=[]):
-        super().__init__(test=test, device=device, jit=jit, batch_size=batch_size, extra_args=extra_args)
-        parser = default_argument_parser()
-        args = parser.parse_args(["--config-file", get_abs_path(variant)])
-        data_cfg = model_zoo.get_config("common/data/coco.py").dataloader
-        cfg = setup(args)
-        self.model = build_model(cfg).to(self.device)
-        if self.test == "train":
-            self.model.train()
-            self.optimizer = build_optimizer(cfg, self.model)
-            # setup train dataset
-            data_cfg.train.dataset.names = "coco_2017_val_100"
-            data_cfg.train.total_batch_size = self.batch_size
-            train_loader = instantiate(data_cfg.train)
-            self.example_inputs = prefetch(itertools.islice(train_loader, 100), self.device)
-        elif self.test == "eval":
-            self.model.eval()
-            # setup eval dataset
-            data_cfg.test.dataset.names = "coco_2017_val_100"
-            data_cfg.test.batch_size = self.batch_size
-            test_loader = instantiate(data_cfg.test)
-            self.example_inputs = prefetch(itertools.islice(test_loader, 100), self.device)
-        cfg.defrost()
-
-    def get_module(self):
-        return self.model, self.example_inputs
-
-    def enable_fp16_half(self):
-        assert self.dargs.precision == "fp16", f"Expected precision fp16, get {self.dargs.precision}"
-        self.model = self.model.half()
-        self.example_inputs = prefetch(self.example_inputs, self.device, self.dargs.precision)
-
-    def train(self):
-        with EventStorage():
-            for _idx, d in enumerate(self.example_inputs):
-                loss_dict = self.model(d)
-                if isinstance(loss_dict, torch.Tensor):
-                    losses = loss_dict
-                    loss_dict = {"total_loss": loss_dict}
-                else:
-                    losses = sum(loss_dict.values())
-                self.optimizer.zero_grad()
-                losses.backward()
-                self.optimizer.step()
-
-    def eval(self) -> Tuple[torch.Tensor]:
-        with torch.no_grad():
-            for _idx, d in enumerate(self.example_inputs):
-                out = self.model(d)
-                # retrieve output tensors
-        outputs = []
-        for item in out:
-            fields = list(map(lambda x: list(x.get_fields().values()), item.values()))
-            for boxes in fields:
-                tensor_box = list(filter(lambda x: isinstance(x, torch.Tensor), boxes))
-                outputs.extend(tensor_box)
-        return tuple(outputs)
+def install_detectron2(model_name, model_dir):
+    check_data_dir()
+    install_model_weights(model_name, model_dir)
+    pip_install_requirements()
