+++ torchbenchmark/util/framework/detectron2/requirements.txt
